+++ add_criterion_min_score_column.sql
@@ -4,13 +4,13 @@
 USE icei_38697196_coecsathesis;
 
 ALTER TABLE `rubric_criteria` 
-ADD COLUMN `min_score` DECIMAL(5,2) NULL DEFAULT 0.00 
+ADD COLUMN `min_score` INT NULL DEFAULT 0 
 COMMENT 'Minimum score for this criterion (used for individual scoring in numerical rubrics)' 
 AFTER `max_score`;
 
 -- Update existing records: set min_score to 0 for existing individual criteria
 UPDATE `rubric_criteria` 
-SET `min_score` = 0.00 
+SET `min_score` = 0 
 WHERE `is_individual` = 1 AND `min_score` IS NULL;
 
 -- Note: Both min_score and max_score define the range for individual criteria
